moved mesh from encoder.py to mesh.py
